requirements.txt
# PowerPoint generation
python-pptx>=0.6.21

# AI/LLM libraries
openai>=1.0.0
anthropic>=0.18.0

# Configuration and utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0

# Image processing (for slide content)
Pillow>=10.0.0

# Logging
loguru>=0.7.0

# CLI
click>=8.0.0
rich>=13.0.0
